==> 00.Load.SCT&h5.ipynb <==
library(dplyr)
library(Matrix)
library(data.table)
library(Seurat)
library(RColorBrewer)
library(cowplot)
library(future)
library(ComplexHeatmap)
library(viridis)
library(SeuratObject)
library(readxl)
library(BayesSpace)
library(SingleCellExperiment)
library(ggplot2)
library(getopt)
library(reshape2)
library(gridExtra)
library(harmony)
library(imager)
library(ArchR)
library(ggplot2)
library(ggrastr)
library(parallel) 
library(ggplot2)
library(cols4all)
library(cowplot)
library("shinyjs")
library("kableExtra")
library("colorblindcheck")
library(tidydr) 
options(warn=-1) 
source("/data/work/AD/bin/batch2/00.code/function.R")
cols <- c(brewer.pal(9, "Set1"),brewer.pal(8,"Set2")[1:8],brewer.pal(12,"Paired")[1:12],brewer.pal(8,"Dark2")[1:8],brewer.pal(8,"Accent"))
suppressMessages(library(tidyverse))
suppressMessages(library(parallel))
suppressMessages(library(Seurat))
# suppressMessages(library(SeuratData))
suppressMessages(library(patchwork))
suppressMessages(library(Matrix))
suppressMessages(library(viridis))
suppressMessages(library(cowplot))
suppressMessages(library(pheatmap))
suppressMessages(library(RColorBrewer))
suppressMessages(library(reshape2))
suppressMessages(library(randomForest))
suppressMessages(library(harmony))
# suppressMessages(library(BPCells))
suppressMessages(library(reticulate))

list.files("/data/users/heyouzhe/online/01_DATA/09_Stereo_matrix/gem/")

chipname <- chipname

setwd("/data/work/AD/bin/batch5_V7/01.SCT&h5.data")

source("/data/work/41.readgem/LoadBGI_Spatial.r")
# dir.create(paste0(args[6],"/01.seurat.SCT"),recursive = TRUE)
# setwd(paste0(args[6],"/01.seurat.SCT"))
seuobj = LoadBGI_Spatial(filename = paste0("/data/users/heyouzhe/online/01_DATA/09_Stereo_matrix/gem/",chipname,".tissue.gem.gz"),cell_mask = F,bin_data = T,UMI_GreyScale_Image = F,area_mask=F,bin_size =100,save_rds=F,pro_name=chipname,save_h5ad=T)

colors <- c(
  "#B22A6D",   # RGB (178, 42, 109)
  "#014F7F",   # RGB (1, 79, 127)
  "#6D5D77",   # RGB (109, 93, 119)
  "#FFB7FF",   # RGB (255, 183, 255)
  "#FF8C53",   # RGB (255, 140, 83)
  "#00555A",   # RGB (0, 85, 90)
  "#01FFFF",   # RGB (1, 255, 255)
  "#924A79",   # RGB (146, 74, 121)
  "#FC97FE",   # RGB (252, 151, 254)
  "#5A8895",   # RGB (90, 136, 149)
  "#0050A0",   # RGB (0, 80, 160)
  "#919BE2",   # RGB (145, 155, 226)
  "#009EBE",   # RGB (0, 158, 190)
  "#059445",   # RGB (5, 148, 69)
  "#74876B",   # RGB (116, 135, 107)
  "#00B4FD"    # RGB (0, 180, 253)
)
best_color <- c("#9571d9", "#99cc31", "#fb8071", "#8cd2c7", "#208b23", "#199a73","#cbeac4", "#6a8e20", "#32a02d", "#a4cee1", "#02bdfe","#08529d", "#108b96", "#fcb462", 
                "#fd7c01", "#fdbc66", "#da5e05", "#8b0201","#E31A1C", "#FDBF6F", "#FF7F00", "#CAB2D6", "#6A3D9A", "#B15928","#FBB4AE", "#B3CDE3", "#BC80BD", "#CCEBC5",
                "#DECBE4", "#FED9A6","#FFFFCC", "#E5D8BD", "#FDDAEC", "#F2F2F2", "#B3E2CD", "#FDCDAC","#CBD5E8", "#F4CAE4", "#E6F5C9", "#FFF2AE", "#F1E2CC", "#CCCCCC",
                "#E41A1C", "#377EB8", "#984EA3", "#FFFF33", "#A65628", "#F781BF","#999999", "#FFED6F", "#66C2A5", "#FC8D62", "#8DA0CB", "#E78AC3","#A6D854", "#FFD92F",
                "#E5C494", "#B3B3B3", "#8DD3C7", "#FFFFB3","#BEBADA", "#FB8072", "#80B1D3", "#FDB462", "#B3DE69", "#FCCDE5","#D9D9D9","#666666")

colsblack <- c(colors,"#FFFF00","#E31A1C","#D7E3FF","#A1CAF1","#C2B280","#999999","#A30059","#FFF300","#F38400",
               "#00FF00","#A020F0","#509DEA","#FF2F80","#FE4945","#FFE4E1","#E25822","#00C2A0","#4FC601","#FFAA92",
               "#FDC086","#FF4A9C","#E68FAC","#F781BF","#8FB0FF","#63FFAC","#1CE6FF","#41AB5D","#0000A6","#FFFF99",
               "#F6A600","#FF1493","#bc321a","#FF90C9","#3B5DFF","#FF34FF","#1E90FF","#F99379","#74C476","#7B4F4B",
               "#DCD300","#363636",best_color)

seuobj <- seuseuobj
seuobj <- NormalizeData(seuobj, normalization.method = "RC") 
seuobj <- SCTransform(seuobj, assay = "Spatial", verbose = FALSE) 
seuobj <- RunPCA(seuobj, assay = "SCT", verbose = FALSE) 
seuobj <- FindNeighbors(seuobj, reduction = "pca", dims = 1:40) 
seuobj <- FindClusters(seuobj, verbose = FALSE,resolution = c(0.5,0.8,1.2,1.5,2)) 
seuobj <- RunUMAP(seuobj, reduction = "pca", dims = 1:40) 

seuobj$coor_y = as.integer(seuobj@images$slice1@coordinates$row)
seuobj$coor_x = as.integer(seuobj@images$slice1@coordinates$col)
coord <- data.frame(stereo_1 = seuobj$coor_x, stereo_2 = seuobj$coor_y)
seuobj[["stereo"]] <- SeuratObject::CreateDimReducObject(embeddings = as.matrix(coord), key = "stereo_", assay = "SCT")
saveRDS(seuobj,paste0(chipname,".bin100.SCT.rds"))

p1 <- list()
p1[[1]]<- ggplot(data=seuobj@meta.data, aes(x=coor_x, y=coor_y, color=SCT_snn_res.0.5)) +  
  geom_point_rast(size=0.1, alpha=1) + 
  scale_color_manual(values=colsblack) + 
  coord_fixed() +
  # geom_point_rast(data=all, aes(x=coor_x, y=coor_y), size=0.2, color='black', stroke=0.0001, alpha=0.1) +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill='white'),
        plot.background=element_rect(fill="white")) + 
  guides(color = guide_legend(override.aes = list(size=1.5))) 

p1[[2]]<- ggplot(data=seuobj@meta.data, aes(x=coor_x, y=coor_y, color=SCT_snn_res.0.8)) +  
  geom_point_rast(size=0.1, alpha=1) + 
  scale_color_manual(values=colsblack) + 
  coord_fixed() +
  # geom_point_rast(data=all, aes(x=coor_x, y=coor_y), size=0.2, color='black', stroke=0.0001, alpha=0.1) +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill='white'),
        plot.background=element_rect(fill="white")) + 
  guides(color = guide_legend(override.aes = list(size=1.5))) 

p1[[3]]<- ggplot(data=seuobj@meta.data, aes(x=coor_x, y=coor_y, color=SCT_snn_res.1.2)) +  
  geom_point_rast(size=0.1, alpha=1) + 
  scale_color_manual(values=colsblack) + 
  coord_fixed() +
  # geom_point_rast(data=all, aes(x=coor_x, y=coor_y), size=0.2, color='black', stroke=0.0001, alpha=0.1) +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill='white'),
        plot.background=element_rect(fill="white")) + 
  guides(color = guide_legend(override.aes = list(size=1.5))) 

p1[[4]]<- ggplot(data=seuobj@meta.data, aes(x=coor_x, y=coor_y, color=SCT_snn_res.1.5)) +  
  geom_point_rast(size=0.1, alpha=1) + 
  scale_color_manual(values=colsblack) + 
  coord_fixed() +
  # geom_point_rast(data=all, aes(x=coor_x, y=coor_y), size=0.2, color='black', stroke=0.0001, alpha=0.1) +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill='white'),
        plot.background=element_rect(fill="white")) + 
  guides(color = guide_legend(override.aes = list(size=1.5))) 

p1[[5]]<- ggplot(data=seuobj@meta.data, aes(x=coor_x, y=coor_y, color=SCT_snn_res.2)) +  
  geom_point_rast(size=0.1, alpha=1) + 
  scale_color_manual(values=colsblack) + 
  coord_fixed() +
  # geom_point_rast(data=all, aes(x=coor_x, y=coor_y), size=0.2, color='black', stroke=0.0001, alpha=0.1) +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill='white'),
        plot.background=element_rect(fill="white")) + 
  guides(color = guide_legend(override.aes = list(size=1.5))) 
# ggsave(paste0(chipname, "_Bin100.thm.pdf"), p1, width=6, height=5)

library(cowplot)
options(warn=-1) 
p_all=plot_grid(plotlist = p1,ncol = 5)
#p_all
ggsave(paste0("/data/work/AD/bin/batch5_V7/01.SCT&h5.data/",chipname,"_SCTclustering.pdf"),p_all,width = 30,height=5)

# obj <- seuobj

# obj$row = obj$coor_y
# obj$col = obj$coor_x

# obj1$row = obj1$coor_y
# obj1$col = obj1$coor_x

# sce <- as.SingleCellExperiment(obj1,assay='SCT')
# set.seed(1234)
# sce <- spatialPreprocess(sce, platform="ST", 
#                          n.PCs=50, n.HVGs=3000, log.normalize=F)



# n_clu=50 
# sce <- spatialCluster(sce, q=n_clu,
#                       platform="ST", 
#                       d=30,
#                       init.method="mclust",
#                       model="t",
#                       gamma=2,
#                       nrep=10000, 
#                       burn.in=100, 
#                       save.chain=TRUE)



# fname3=paste0(chipname,".clusterPlot",n_clu,".pdf")
# pdf(fname3)
# p = clusterPlot(sce,label = "spatial.cluster",color=NA,palette=colsblack[1:n_clu])
# print(p)
# dev.off()
# meta = colData(sce)
# fname6<-paste0(chipname,'.BayesSpace.clustering',n_clu,'.rds')
# meta = meta[as.matrix(rownames(obj@meta.data)),]
# saveRDS(meta,fname6)
