==> 01.Make.h5.task (1).ipynb <==
import stereo as st
import warnings
warnings.filterwarnings('ignore')
import os
new_path = '/data/work/AD/bin/batch5/01.H5.data'  
os.chdir(new_path)

# 定义taskid和SN列表
taskid_list = [
    'WT202410120001744', 'WT202410120001743', 'WT202410120000008',
    'WT202410120000007', 'WT202410120000005', 'WT202410120000004', 'WT202410120000003',
    'WT202410120000002', 'WT202410120000001', 'WT202410110005948', 'WT202410110005947',
    'WT202410110005945', 'WT202410110005944'
]

SN_list = [
    'D04305A6', 'D04305A4', 'C04595F1', 'D03556C4', 'D03556D4', 'D03556D6',
    'D03556E4', 'D03556E6', 'C04595E2', 'D03556F4', 'D03556F6', 'D04303A6', 'D03556E2'
]


for taskid, SN in zip(taskid_list, SN_list):
    data_path = f'/data/input/{taskid}/Copy-SAW-ST-V8/result/{SN}/outs/feature_expression/{SN}.tissue.gef'
    print(data_path)
    data = st.io.read_gef(file_path=data_path, bin_size=100)
    adata = st.io.stereo_to_anndata(data,flavor='seurat',output=f'{SN}.h5ad')

taskid

data_path = f'/data/input/{taskid}/Copy-SAW-ST-V8(1)/result/{SN}/outs/feature_expression/{SN}.tissue.gef'
print(data_path)
data = st.io.read_gef(file_path=data_path, bin_size=100)
adata = st.io.stereo_to_anndata(data,flavor='seurat',output=f'{SN}.h5ad')

st.__version__

taskid = taskid
SN = SN

data_path = f'/data/input/{taskid}/Copy-SAW-ST-V8/result/{SN}/outs/feature_expression/{SN}.tissue.gef'
st.io.read_gef_info(data_path)
data = st.io.read_gef(file_path=data_path, bin_size=100)
adata = st.io.stereo_to_anndata(data,flavor='seurat',output=f'{SN}.h5ad')
