==> 02.Make.matrix.task.ipynb <==
import pandas as pd
import scanpy as sc
import os
new_path = '/data/work/AD/bin/batch5/02.Matrix.data'  
os.chdir(new_path)

SN_list = [
    'D04305A6', 'D04305A4', 'C04595F1', 'D03556C4', 'D03556D4', 'D03556D6',
    'D03556E4', 'D03556E6', 'C04595E2', 'D03556F4', 'D03556F6', 'D04303A6', 'D03556E2'
]


for SN in SN_list:
    data_path = f'/data/work/AD/bin/batch5/01.H5.data/{SN}.h5ad'
    print(data_path)
    adata = sc.read_h5ad(data_path)
    matrix = adata.X
    dense_matrix = pd.DataFrame(matrix.todense())
    gene_names_df = adata.var[['real_gene_name']]
    gene_names_df_reset = gene_names_df.reset_index()
    gene_names_df_reset.to_csv(f'{SN}.features.csv', index=False)
    dense_matrix.to_parquet(f'{SN}.matrix.parquet')
    xy = adata.obs[['x', 'y']]
    xy.to_csv(f'{SN}_coordinates.csv', index=False)

for SN in SN_list:
    data_path = f'/data/work/AD/bin/batch5/01.H5.data/{SN}.h5ad'
    print(f'Processing: {data_path}')
    
    # 读取 H5AD 文件
    adata = sc.read_h5ad(data_path)
    
    # 提取表达矩阵 (dense matrix) 并保留行名
    matrix = adata.X
    dense_matrix = pd.DataFrame(matrix.todense(), index=adata.obs_names)  # 将 obs_names 作为行名
    
    # 保存基因名
    gene_names_df = adata.var[['real_gene_name']]
    gene_names_df_reset = gene_names_df.reset_index()
    gene_names_df_reset.to_csv(f'{SN}.features.csv', index=False)
    dense_matrix['row_names'] = dense_matrix.index
    dense_matrix.to_parquet(f'{SN}.cellid.parquet', index=False)
    # 保存表达矩阵为 parquet 文件，并保留行名
    dense_matrix.to_parquet(f'{SN}.matrix.parquet')
    
    # 保存坐标数据
    xy = adata.obs[['x', 'y']]
    xy.to_csv(f'{SN}_coordinates.csv', index=False)


for SN in SN_list:
    data_path = f'/data/work/AD/bin/batch5/01.H5.data/{SN}.h5ad'
    print(f'Processing: {data_path}')
    
    # 读取 H5AD 文件
    adata = sc.read_h5ad(data_path)
    
    # 提取表达矩阵 (dense matrix) 并保留行名
    matrix = adata.X
    dense_matrix = pd.DataFrame(matrix.todense(), index=adata.obs_names)  # 将 obs_names 作为行名
    dense_matrix['row_names'] = dense_matrix.index
    dense_matrix.to_parquet(f'{SN}.cellid.parquet', index=False)


dense_matrix

len(set(adata.var.real_gene_name))

# 查看 adata.X 的列名
print(adata.var_names)

adata.var.real_gene_name
