==> 03.Make.seurat.task.ipynb <==
library(data.table)
library(arrow)
library(ArchR)
library(Seurat)
library(dplyr)
library(Matrix)
library(RColorBrewer)
library(cowplot)
library(future)
library(viridis)
library(SeuratObject)
library(BayesSpace)
library(SingleCellExperiment)
library(ggplot2)
library(getopt)
library(reshape2)
library(gridExtra)
library(harmony)
library(imager)
library(ggrastr)
library(parallel) 
library(cols4all)
library(cowplot)
library("shinyjs")
library("kableExtra")
library("colorblindcheck")
library(tidydr) 
options(warn=-1) 
suppressMessages(library(tidyverse))
suppressMessages(library(parallel))
suppressMessages(library(Seurat))
# suppressMessages(library(SeuratData))
suppressMessages(library(patchwork))
suppressMessages(library(Matrix))
suppressMessages(library(viridis))
suppressMessages(library(cowplot))
suppressMessages(library(pheatmap))
suppressMessages(library(RColorBrewer))
suppressMessages(library(reshape2))
suppressMessages(library(randomForest))
suppressMessages(library(harmony))
# suppressMessages(library(BPCells))
suppressMessages(library(reticulate))

colors <- c(
  "#B22A6D",   # RGB (178, 42, 109)
  "#014F7F",   # RGB (1, 79, 127)
  "#6D5D77",   # RGB (109, 93, 119)
  "#FFB7FF",   # RGB (255, 183, 255)
  "#FF8C53",   # RGB (255, 140, 83)
  "#00555A",   # RGB (0, 85, 90)
  "#01FFFF",   # RGB (1, 255, 255)
  "#924A79",   # RGB (146, 74, 121)
  "#FC97FE",   # RGB (252, 151, 254)
  "#5A8895",   # RGB (90, 136, 149)
  "#0050A0",   # RGB (0, 80, 160)
  "#919BE2",   # RGB (145, 155, 226)
  "#009EBE",   # RGB (0, 158, 190)
  "#059445",   # RGB (5, 148, 69)
  "#74876B",   # RGB (116, 135, 107)
  "#00B4FD"    # RGB (0, 180, 253)
)
best_color <- c("#9571d9", "#99cc31", "#fb8071", "#8cd2c7", "#208b23", "#199a73","#cbeac4", "#6a8e20", "#32a02d", "#a4cee1", "#02bdfe","#08529d", "#108b96", "#fcb462", 
                "#fd7c01", "#fdbc66", "#da5e05", "#8b0201","#E31A1C", "#FDBF6F", "#FF7F00", "#CAB2D6", "#6A3D9A", "#B15928","#FBB4AE", "#B3CDE3", "#BC80BD", "#CCEBC5",
                "#DECBE4", "#FED9A6","#FFFFCC", "#E5D8BD", "#FDDAEC", "#F2F2F2", "#B3E2CD", "#FDCDAC","#CBD5E8", "#F4CAE4", "#E6F5C9", "#FFF2AE", "#F1E2CC", "#CCCCCC",
                "#E41A1C", "#377EB8", "#984EA3", "#FFFF33", "#A65628", "#F781BF","#999999", "#FFED6F", "#66C2A5", "#FC8D62", "#8DA0CB", "#E78AC3","#A6D854", "#FFD92F",
                "#E5C494", "#B3B3B3", "#8DD3C7", "#FFFFB3","#BEBADA", "#FB8072", "#80B1D3", "#FDB462", "#B3DE69", "#FCCDE5","#D9D9D9","#666666")

colsblack <- c(colors,"#FFFF00","#E31A1C","#D7E3FF","#A1CAF1","#C2B280","#999999","#A30059","#FFF300","#F38400",
               "#00FF00","#A020F0","#509DEA","#FF2F80","#FE4945","#FFE4E1","#E25822","#00C2A0","#4FC601","#FFAA92",
               "#FDC086","#FF4A9C","#E68FAC","#F781BF","#8FB0FF","#63FFAC","#1CE6FF","#41AB5D","#0000A6","#FFFF99",
               "#F6A600","#FF1493","#bc321a","#FF90C9","#3B5DFF","#FF34FF","#1E90FF","#F99379","#74C476","#7B4F4B",
               "#DCD300","#363636",best_color)

# obj <- readRDS("/data/work/AD/bin/batch5/04.SCT&h5.data/D03556C4.bin100.SCT.rds")

# install.packages("/data/work/ADSC/01.data/Astro/TheBestColors_0.1.0.tar.gz", repos = NULL, type = "source")

# library(TheBestColors)

# map <- Best100(90,256)
# # map <- Best100(76,256)
# mytheme <- theme_void()

# df <- obj@meta.data
# gene <- "RGS14"
# df$expr <- obj@assays$SCT@data[gene,]
# options(repr.plot.width=5, repr.plot.height=5)
# ggplot(data = df, aes(x = coor_x, y = coor_y)) +
#     geom_point_rast(aes(color = expr),
#                     size = 0.5, 
#                     alpha =1)+
#     # geom_point_rast(data=line,aes(x=coor_x,y=coor_y),size=0.2,color='black',stroke = 0.0001,alpha=0.3) + #画线
#     mytheme+
#     theme_dr(xlength = 0.2,ylength = 0.2, 
#              arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
#     theme(panel.grid = element_blank())+
#     scale_color_gradientn(colors = map) +
#     labs(title = gene)+
#     theme(plot.title = element_text(size=15,hjust = 0.5,face ="italic"))
#   #   theme(
#   #   panel.background = element_rect(fill = "black"),  # 设置图形区域背景为黑色
#   #   plot.background = element_rect(fill = "black"),   # 设置整个图形的背景为黑色
#   #   # panel.grid = element_line(color = "gray"),        # 网格线的颜色
#   #   # axis.text = element_text(color = "white"),        # 坐标轴文字的颜色
#   #   plot.title = element_text(color = "white",face ="italic" )         # 坐标轴标题的颜色
#   # )

# df <- obj@meta.data
# gene <- "RGS5"
# df$expr <- obj@assays$SCT@data[gene,]
# options(repr.plot.width=5, repr.plot.height=5)
# ggplot(data = df, aes(x = coor_x, y = coor_y)) +
#     geom_point_rast(aes(color = expr),
#                     size = 0.5, 
#                     alpha =1)+
#     # geom_point_rast(data=line,aes(x=coor_x,y=coor_y),size=0.2,color='black',stroke = 0.0001,alpha=0.3) + #画线
#     mytheme+
#     theme_dr(xlength = 0.2,ylength = 0.2, 
#              arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
#     theme(panel.grid = element_blank())+
#     scale_color_gradientn(colors = map) +
#     labs(title = gene)+
#     theme(plot.title = element_text(size=15,hjust = 0.5,face ="italic"))
#   #   theme(
#   #   panel.background = element_rect(fill = "black"),  # 设置图形区域背景为黑色
#   #   plot.background = element_rect(fill = "black"),   # 设置整个图形的背景为黑色
#   #   # panel.grid = element_line(color = "gray"),        # 网格线的颜色
#   #   # axis.text = element_text(color = "white"),        # 坐标轴文字的颜色
#   #   plot.title = element_text(color = "white",face ="italic" )         # 坐标轴标题的颜色
#   # )

SN_list = c('D04305A6', 'D04305A4', 'C04595F1', 'D03556C4', 'D03556D4', 'D03556D6',
            'D03556E4', 'D03556E6', 'C04595E2', 'D03556F4', 'D03556F6', 'D04303A6', 'D03556E2')

i=1
features <- fread(paste0("/data/work/AD/bin/batch5/02.Matrix.data/",SN_list[i],".features.csv"))
  locs <- fread(paste0("/data/work/AD/bin/batch5/02.Matrix.data/",SN_list[i],"_coordinates.csv"))
  expr <- as.data.frame(read_parquet(paste0("/data/work/AD/bin/batch5/02.Matrix.data/",SN_list[i],".cellid.parquet")))
  rownames(expr) <- expr$row_names
  expr <- expr[,-ncol(expr)]
  duplicates.genes <- unique(features$real_gene_name[which(duplicated(features$real_gene_name))])

duplicates.genes

for (i in 1:length(SN_list)){
  features <- fread(paste0("/data/work/AD/bin/batch5/02.Matrix.data/",SN_list[i],".features.csv"))
  locs <- fread(paste0("/data/work/AD/bin/batch5/02.Matrix.data/",SN_list[i],"_coordinates.csv"))
  expr <- as.data.frame(read_parquet(paste0("/data/work/AD/bin/batch5/02.Matrix.data/",SN_list[i],".cellid.parquet")))
  rownames(expr) <- expr$row_names
  expr <- expr[,-ncol(expr)]
  duplicates.genes <- unique(features$real_gene_name[which(duplicated(features$real_gene_name))])
  print(length(duplicates.genes))
  if (length(duplicates.genes) > 0 ) {
    index1 <- which(features$real_gene_name %in% duplicates.genes)
    index2 <- which(features$real_gene_name %ni% duplicates.genes)
    df1 <- expr[,index1]
    df2 <- expr[,index2]
    colnames(df2) <- features$real_gene_name[index2]
    df.list <- list()
    index.gene <- features$real_gene_name[index1]
    for(z in 1:length(duplicates.genes)){
      gene <- duplicates.genes[z]
      df3 <- df1[,which(index.gene == gene)]
      df3 <- rowMeans(df3)
      df.list[[z]] <- df3
    }
    df4 <- do.call(cbind,df.list)
    colnames(df4) <- duplicates.genes
    df4 <- round(df4,0)
    df <- cbind(df2,df4)
  } else {
    df <- expr
    colnames(df) <- features$real_gene_name
  }
  df <- as.data.frame(t(df))
  seuobj <- CreateSeuratObject(counts = df, project = "Spatial",assay = "Spatial")
  rownames(locs) <- colnames(df)
  colnames(locs) <- c("coor_x","coor_y")
  seuobj <- AddMetaData(seuobj, metadata = locs)
  # coord <- data.frame(stereo_1 = seuobj$coor_x, stereo_2 = seuobj$coor_y)
  # seuobj[["stereo"]] <- SeuratObject::CreateDimReducObject(embeddings = as.matrix(coord), key = "stereo_", assay = "SCT")
  saveRDS(seuobj,file = paste0("/data/work/AD/bin/batch5/03.Raw.rds.data/",SN_list[i],".bin100.raw.rds"))
  
}
