==> 08.Marker.print.ipynb <==
library(dplyr)
library(Matrix)
library(data.table)
library(Seurat)
library(RColorBrewer)
library(cowplot)
library(future)
library(ComplexHeatmap)
library(viridis)
library(SeuratObject)
library(readxl)
library(BayesSpace)
library(SingleCellExperiment)
library(ggplot2)
library(getopt)
library(reshape2)
library(gridExtra)
library(harmony)
library(imager)
library(ArchR)
library(ggplot2)
library(ggrastr)
library(parallel) 
library(ggplot2)
library(cols4all)
library(cowplot)
library("shinyjs")
library("kableExtra")
library("colorblindcheck")
library(tidydr) 
options(warn=-1) 
# source("/data/work/AD/bin/batch2/00.code/function.R")
cols <- c(brewer.pal(9, "Set1"),brewer.pal(8,"Set2")[1:8],brewer.pal(12,"Paired")[1:12],brewer.pal(8,"Dark2")[1:8],brewer.pal(8,"Accent"))
suppressMessages(library(tidyverse))
suppressMessages(library(parallel))
suppressMessages(library(Seurat))
# suppressMessages(library(SeuratData))
suppressMessages(library(patchwork))
suppressMessages(library(Matrix))
suppressMessages(library(viridis))
suppressMessages(library(cowplot))
suppressMessages(library(pheatmap))
suppressMessages(library(RColorBrewer))
suppressMessages(library(reshape2))
suppressMessages(library(randomForest))
suppressMessages(library(harmony))
# suppressMessages(library(BPCells))
suppressMessages(library(reticulate))

colors <- c(
  "#B22A6D",   # RGB (178, 42, 109)
  "#014F7F",   # RGB (1, 79, 127)
  "#6D5D77",   # RGB (109, 93, 119)
  "#FFB7FF",   # RGB (255, 183, 255)
  "#FF8C53",   # RGB (255, 140, 83)
  "#00555A",   # RGB (0, 85, 90)
  "#01FFFF",   # RGB (1, 255, 255)
  "#924A79",   # RGB (146, 74, 121)
  "#FC97FE",   # RGB (252, 151, 254)
  "#5A8895",   # RGB (90, 136, 149)
  "#0050A0",   # RGB (0, 80, 160)
  "#919BE2",   # RGB (145, 155, 226)
  "#009EBE",   # RGB (0, 158, 190)
  "#059445",   # RGB (5, 148, 69)
  "#74876B",   # RGB (116, 135, 107)
  "#00B4FD"    # RGB (0, 180, 253)
)
best_color <- c("#9571d9", "#99cc31", "#fb8071", "#8cd2c7", "#208b23", "#199a73","#cbeac4", "#6a8e20", "#32a02d", "#a4cee1", "#02bdfe","#08529d", "#108b96", "#fcb462", 
                "#fd7c01", "#fdbc66", "#da5e05", "#8b0201","#E31A1C", "#FDBF6F", "#FF7F00", "#CAB2D6", "#6A3D9A", "#B15928","#FBB4AE", "#B3CDE3", "#BC80BD", "#CCEBC5",
                "#DECBE4", "#FED9A6","#FFFFCC", "#E5D8BD", "#FDDAEC", "#F2F2F2", "#B3E2CD", "#FDCDAC","#CBD5E8", "#F4CAE4", "#E6F5C9", "#FFF2AE", "#F1E2CC", "#CCCCCC",
                "#E41A1C", "#377EB8", "#984EA3", "#FFFF33", "#A65628", "#F781BF","#999999", "#FFED6F", "#66C2A5", "#FC8D62", "#8DA0CB", "#E78AC3","#A6D854", "#FFD92F",
                "#E5C494", "#B3B3B3", "#8DD3C7", "#FFFFB3","#BEBADA", "#FB8072", "#80B1D3", "#FDB462", "#B3DE69", "#FCCDE5","#D9D9D9","#666666")

colsblack <- c(colors,"#FFFF00","#E31A1C","#D7E3FF","#A1CAF1","#C2B280","#999999","#A30059","#FFF300","#F38400",
               "#00FF00","#A020F0","#509DEA","#FF2F80","#FE4945","#FFE4E1","#E25822","#00C2A0","#4FC601","#FFAA92",
               "#FDC086","#FF4A9C","#E68FAC","#F781BF","#8FB0FF","#63FFAC","#1CE6FF","#41AB5D","#0000A6","#FFFF99",
               "#F6A600","#FF1493","#bc321a","#FF90C9","#3B5DFF","#FF34FF","#1E90FF","#F99379","#74C476","#7B4F4B",
               "#DCD300","#363636",best_color)

install.packages("/data/work/ADSC/01.data/Astro/TheBestColors_0.1.0.tar.gz", repos = NULL, type = "source")
library(TheBestColors)
map <- Best100(90,256)
# map <- Best100(76,256)
mytheme <- theme_void()

setwd("/data/work/AD/bin/batch5_V7/11.Bin.Registration32.data")

chip.name <- c(   
    'Con1.1',
    'Con1.2',
    'Con2.1',
    'Con2.2',
    'Con3.1',
    'Con3.2',
    'Con4.1',
    'Con4.2',
    'Con5.1',
    'Con5.2',
    'Con6.1',
    'Con6.2',
    'Con7.1',
    'Con7.2',
    'Con8.1',
    'Con8.2',
    
    'AD1.1',
    'AD1.2',
    'AD2.1',
    'AD2.2',
    'AD3.1',
    'AD3.2',
    'AD4.1',
    'AD4.2',
    'AD5.1',
    'AD5.2',
    'AD6.1',
    'AD6.2',
    'AD7.1',
    'AD7.2',
    'AD8.1',
    'AD8.2'
)

load("/data/work/AD/bin/batch5_V7/11.Bin.Registration32.data/metainfo.Rdata")

SN_list <- c(
    'D03556D6',
    'D03556E2',
    'D04305A4',
    'D04305A6',
    'D03556F4',
    'D03556F6',
    'C04595E2',
    'C04595F1',

    'D03556C4',
    'D03556D4',
    'D04303A6',
    'D04303D1',
    'D03556E4',
    'D03556E6',
    'D04305A2',
    'D04305C6')

useid <- c(
    'Con5.1',
    'Con5.2',
    'Con6.1',
    'Con6.2',
    'Con7.1',
    'Con7.2',
    'Con8.1',
    'Con8.2',
    
    'AD5.1',
    'AD5.2',
    'AD6.1',
    'AD6.2',
    'AD7.1',
    'AD7.2',
    'AD8.1',
    'AD8.2'
)

for (i in 1:16){
    line.list1234[[useid[i]]] <- line.list5[[SN_list[i]]]
}


i <- 8
obj <- readRDS(paste0("/data/work/AD/bin/batch5_V7/07.Bin100.lasso.area.data/",chip.name[i],".bin100.SCT.lasso.rds"))


options(repr.plot.width=5, repr.plot.height=5)
gene <- "PLP1"
df <- obj@meta.data
df$expr <- obj@assays$SCT@data[gene,]
p1 <- ggplot(data = df, aes(x = rx, y = ry)) +
    geom_point_rast(aes(color = expr),
                    size = 1,
                    stroke = 0.0001,
                    alpha =1)+
    geom_point_rast(data=line.list1234[[chip.name[i]]],aes(x=rx,y=ry),size=0.2,color='black',stroke = 0.0001,alpha=1) + #画线
    mytheme+
    theme_dr(xlength = 0.2,ylength = 0.2, 
             arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
    theme(panel.grid = element_blank(),
         legend.position="none")+
    scale_color_gradientn(colors = map) +
    labs(title = paste0(gene,"--",chip.name[i]))+
    theme(plot.title = element_text(size=10,hjust = 0.5,face ="italic"))
ggsave("CX3CR1.pdf",p1,width=6,height=6.5)

obj.list <- list()
for (i in 1:length(chip.name)){
    obj.list[[chip.name[i]]] <- readRDS(paste0("/data/work/AD/bin/batch5_V7/07.Bin100.lasso.area.data/",chip.name[i],".bin100.SCT.lasso.rds"))
    obj.list[[chip.name[i]]]$chip <- chip.name[i]
}

###B
# genes <- unique(classMarker$gene)
gene <- "PLPPR4"
p <- list()
df.list <- list()
for (i in 1:32){
  df.list[[i]] <- obj.list[[chip.name[i]]]@meta.data
  df.list[[i]]$expr <- obj.list[[chip.name[i]]]@assays$SCT@data[gene,]
  p[[i]] <- ggplot(data = df.list[[i]], aes(x = rx, y = ry)) +
    geom_point_rast(aes(color = expr),
                    size = 0.2,
                    stroke = 0.0001,
                    alpha =1)+
    geom_point_rast(data=line.list1234[[chip.name[i]]],aes(x=rx,y=ry),size=0.2,color='black',stroke = 0.0001,alpha=0.3) + #画线
    mytheme+
    theme_dr(xlength = 0.2,ylength = 0.2, 
             arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
    theme(panel.grid = element_blank(),
         legend.position="none")+
    scale_color_gradientn(colors = map) +
    labs(title = paste0(gene,"--",chip.name[i]))+
    theme(plot.title = element_text(size=10,hjust = 0.5,face ="italic"))
}


p_all=plot_grid(plotlist =p,ncol = 16)
# options(repr.plot.width=48, repr.plot.height=7)
# p_all
ggsave(paste0(gene,".featureplot.pdf"),p_all,width=24,height=3.5)

###B
# genes <- unique(classMarker$gene)
gene <- "RGS14"
p <- list()
df.list <- list()
for (i in 1:32){
  df.list[[i]] <- obj.list[[chip.name[i]]]@meta.data
  df.list[[i]]$expr <- obj.list[[chip.name[i]]]@assays$SCT@data[gene,]
  p[[i]] <- ggplot(data = df.list[[i]], aes(x = rx, y = ry)) +
    geom_point_rast(aes(color = expr),
                    size = 0.2,
                    stroke = 0.0001,
                    alpha =1)+
    geom_point_rast(data=line.list1234[[chip.name[i]]],aes(x=rx,y=ry),size=0.2,color='black',stroke = 0.0001,alpha=0.3) + #画线
    mytheme+
    theme_dr(xlength = 0.2,ylength = 0.2, 
             arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
    theme(panel.grid = element_blank(),
         legend.position="none")+
    scale_color_gradientn(colors = map) +
    labs(title = paste0(gene,"--",chip.name[i]))+
    theme(plot.title = element_text(size=15,hjust = 0.5,face ="italic"))
}


p_all=plot_grid(plotlist =p,ncol = 16)
# options(repr.plot.width=48, repr.plot.height=7)
# p_all
ggsave(paste0(gene,".featureplot.pdf"),p_all,width=24,height=3.5)

###B
# genes <- unique(classMarker$gene)
gene <- "HOPX"
p <- list()
df.list <- list()
for (i in 1:32){
  df.list[[i]] <- obj.list[[chip.name[i]]]@meta.data
  df.list[[i]]$expr <- obj.list[[chip.name[i]]]@assays$SCT@data[gene,]
  p[[i]] <- ggplot(data = df.list[[i]], aes(x = rx, y = ry)) +
    geom_point_rast(aes(color = expr),
                    size = 0.2,
                    stroke = 0.0001,
                    alpha =1)+
    geom_point_rast(data=line.list1234[[chip.name[i]]],aes(x=rx,y=ry),size=0.2,color='black',stroke = 0.0001,alpha=0.3) + #画线
    mytheme+
    theme_dr(xlength = 0.2,ylength = 0.2, 
             arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
    theme(panel.grid = element_blank(),
         legend.position="none")+
    scale_color_gradientn(colors = map) +
    labs(title = paste0(gene,"--",chip.name[i]))+
    theme(plot.title = element_text(size=15,hjust = 0.5,face ="italic"))
}


p_all=plot_grid(plotlist =p,ncol = 16)
# options(repr.plot.width=48, repr.plot.height=7)
# p_all
ggsave(paste0(gene,".featureplot.pdf"),p_all,width=24,height=3.5)


###B
# genes <- unique(classMarker$gene)
gene <- "OGFRL1"
p <- list()
df.list <- list()
for (i in 1:32){
  df.list[[i]] <- obj.list[[chip.name[i]]]@meta.data
  df.list[[i]]$expr <- obj.list[[chip.name[i]]]@assays$SCT@data[gene,]
  p[[i]] <- ggplot(data = df.list[[i]], aes(x = rx, y = ry)) +
    geom_point_rast(aes(color = expr),
                    size = 0.2,
                    stroke = 0.0001,
                    alpha =1)+
    geom_point_rast(data=line.list1234[[chip.name[i]]],aes(x=rx,y=ry),size=0.2,color='black',stroke = 0.0001,alpha=0.3) + #画线
    mytheme+
    theme_dr(xlength = 0.2,ylength = 0.2, 
             arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
    theme(panel.grid = element_blank(),
         legend.position="none")+
    scale_color_gradientn(colors = map) +
    labs(title = paste0(gene,"--",chip.name[i]))+
    theme(plot.title = element_text(size=15,hjust = 0.5,face ="italic"))
}


p_all=plot_grid(plotlist =p,ncol = 16)
# options(repr.plot.width=48, repr.plot.height=7)
# p_all
ggsave(paste0(gene,".featureplot.pdf"),p_all,width=24,height=3.5)


###B
# genes <- unique(classMarker$gene)
gene <- "SEMA5A"
p <- list()
df.list <- list()
for (i in 1:32){
  df.list[[i]] <- obj.list[[chip.name[i]]]@meta.data
  df.list[[i]]$expr <- obj.list[[chip.name[i]]]@assays$SCT@data[gene,]
  p[[i]] <- ggplot(data = df.list[[i]], aes(x = rx, y = ry)) +
    geom_point_rast(aes(color = expr),
                    size = 0.2,
                    stroke = 0.0001,
                    alpha =1)+
    geom_point_rast(data=line.list1234[[chip.name[i]]],aes(x=rx,y=ry),size=0.2,color='black',stroke = 0.0001,alpha=0.3) + #画线
    mytheme+
    theme_dr(xlength = 0.2,ylength = 0.2, 
             arrow = grid::arrow(length = unit(0.1, "inches"), ends = 'last', type = "closed")) + 
    theme(panel.grid = element_blank(),
         legend.position="none")+
    scale_color_gradientn(colors = map) +
    labs(title = paste0(gene,"--",chip.name[i]))+
    theme(plot.title = element_text(size=15,hjust = 0.5,face ="italic"))
}


p_all=plot_grid(plotlist =p,ncol = 16)
# options(repr.plot.width=48, repr.plot.height=7)
# p_all
ggsave(paste0(gene,".featureplot.pdf"),p_all,width=24,height=3.5)
